# backend/static/openapi.yaml
openapi: "3.0.3"
info:
  title: Auth-as-a-Service API
  description: |
    Context-aware multi-factor authentication service.
    External applications authenticate via API keys and receive JWT tokens.

    ## Authentication
    - **API Key**: Pass `X-API-Key` header on every `/api/v1/*` request.
    - **JWT Bearer**: After login, use `Authorization: Bearer <access_token>` for protected endpoints.
    - **Admin**: Pass `X-Admin-Secret` header for `/api/admin/*` management endpoints.
  version: "1.0.0"
  contact:
    name: Security Verification System
servers:
  - url: http://localhost:5000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (starts with sk_live_)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret for management endpoints
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        has_face:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/User"
    LoginSteps:
      type: object
      properties:
        password:
          type: boolean
        otp:
          type: boolean
        face:
          type: boolean

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  user_id:
                    type: integer
                  message:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- v1: Login flow ------------------------------------------------------
  /api/v1/login/start:
    post:
      tags: [Auth]
      summary: Start multi-step login
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                geo:
                  type: object
                  description: Optional geolocation override
      responses:
        "200":
          description: Login started
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_token:
                    type: string
                  steps:
                    $ref: "#/components/schemas/LoginSteps"
                  verification_level:
                    type: integer
                  context:
                    type: object
  /api/v1/login/verify-password:
    post:
      tags: [Auth]
      summary: Verify password
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_token, password]
              properties:
                session_token:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Password verified
        "401":
          description: Invalid password
  /api/v1/login/verify-otp:
    post:
      tags: [Auth]
      summary: Verify email OTP
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_token, otp]
              properties:
                session_token:
                  type: string
                otp:
                  type: string
      responses:
        "200":
          description: OTP verified
        "401":
          description: Invalid OTP
  /api/v1/login/verify-face:
    post:
      tags: [Auth]
      summary: Verify face recognition
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_token, image]
              properties:
                session_token:
                  type: string
                image:
                  type: string
                  description: Base64-encoded face image
      responses:
        "200":
          description: Face verified
        "401":
          description: Face verification failed
  /api/v1/login/complete:
    post:
      tags: [Auth]
      summary: Complete login and receive JWT tokens
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_token]
              properties:
                session_token:
                  type: string
      responses:
        "200":
          description: Login complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"

  # ---- v1: Token refresh ---------------------------------------------------
  /api/v1/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer

  # ---- v1: Protected user info ---------------------------------------------
  /api/v1/user/me:
    get:
      tags: [Auth]
      summary: Get current user info from JWT
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"

  # ---- v1: Face setup ------------------------------------------------------
  /api/v1/setup/face/multi:
    post:
      tags: [Face]
      summary: Register face from multiple poses
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, images]
              properties:
                user_id:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
                  description: Array of base64-encoded face images
      responses:
        "200":
          description: Face registered
        "400":
          description: No face detected
  /api/v1/check-face:
    post:
      tags: [Face]
      summary: Check if a face is detected in an image
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  description: Base64-encoded image
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                type: object
                properties:
                  face_detected:
                    type: boolean

  # ---- v1: Context ---------------------------------------------------------
  /api/v1/verify-context:
    post:
      tags: [Context]
      summary: Check risk level for a given user context
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                geo:
                  type: object
      responses:
        "200":
          description: Risk assessment
          content:
            application/json:
              schema:
                type: object
                properties:
                  verification_level:
                    type: integer
                  steps:
                    $ref: "#/components/schemas/LoginSteps"
                  context:
                    type: object